+++ src/test/java/pages/business/common/BusinessBaseTestClass.java
@@ -24,6 +24,7 @@ import pages.business.manager.BusinessManagerAddCreatorPage;
 import pages.business.manager.BusinessManagerAddEmployeePage;
 import pages.business.manager.BusinessManagerDeleteCreatorPage;
 import pages.business.manager.BusinessManagerDeleteEmployeePage;
+import pages.business.manager.BusinessManagerLanguagePage;
 import pages.business.manager.BusinessManagerLoginPage;
 import pages.business.manager.BusinessManagerSettingsPage;
 import pages.business.manager.BusinessManagerSignUpPage;
@@ -46,6 +47,7 @@ public class BusinessBaseTestClass {
     protected BusinessManagerAddEmployeePage businessManagerAddEmployeePage;
     protected BusinessManagerDeleteCreatorPage businessManagerDeleteCreatorPage;
     protected BusinessManagerDeleteEmployeePage businessManagerDeleteEmployeePage;
+    protected BusinessManagerLanguagePage businessManagerLanguagePage;
     protected BusinessManagerSettingsPage businessManagerSettingsPage;
     protected BusinessEmployeeSignUpPage businessEmployeeSignUpPage;
     protected BusinessEmployeeLoginPage businessEmployeeLoginPage;
@@ -79,6 +81,7 @@ public class BusinessBaseTestClass {
         businessManagerAddEmployeePage = new BusinessManagerAddEmployeePage(page);
         businessManagerDeleteCreatorPage = new BusinessManagerDeleteCreatorPage(page);
         businessManagerDeleteEmployeePage = new BusinessManagerDeleteEmployeePage(page);
+        businessManagerLanguagePage = new BusinessManagerLanguagePage(page);
         businessManagerSettingsPage = new BusinessManagerSettingsPage(page);
         businessEmployeeSignUpPage = new BusinessEmployeeSignUpPage(page);
         businessEmployeeLoginPage = new BusinessEmployeeLoginPage(page);